requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
python-dotenv
apscheduler
jinja2
aiosmtplib
email-validator
python-multipart
werkzeug
cryptography
cron-descriptor
itsdangerous
bcrypt==3.2.2
passlib[bcrypt]
fastapi-limiter
redis
python-jose[cryptography]